# apps/api/src/main.rs
mod account_deletion;
mod administration;
pub mod api;
mod application;
pub mod authorization;
mod federation;
mod moderation;
mod observability;
mod operators;
mod pagination;
mod posts;
mod profile;
mod registration;
mod relationships;
mod reports;
pub mod security;
mod sessions;
mod social;

use axum::{
    Router,
    extract::{DefaultBodyLimit, State},
    http::StatusCode,
    middleware,
    routing::get,
};
use miz_api::infrastructure;
use security::SecurityState;
use tower::ServiceBuilder;
use tower_http::{
    request_id::{PropagateRequestIdLayer, SetRequestIdLayer},
    trace::TraceLayer,
};
use tracing_subscriber::EnvFilter;

#[tokio::main]
async fn main() {
    tracing_subscriber::fmt()
        .json()
        .with_env_filter(
            EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new("info")),
        )
        .init();
    let database_url = std::env::var("DATABASE_URL").expect("DATABASE_URL must be set");
    let pool = infrastructure::database(&database_url)
        .await
        .expect("database connection and migrations must succeed");
    let cursor_signing_key = std::env::var("CURSOR_SIGNING_KEY")
        .expect("CURSOR_SIGNING_KEY must be set")
        .into_bytes();
    assert!(
        cursor_signing_key.len() >= 32,
        "CURSOR_SIGNING_KEY must contain at least 32 bytes"
    );
    let operator_mfa_secret = std::env::var("OPERATOR_MFA_ENCRYPTION_KEY")
        .expect("OPERATOR_MFA_ENCRYPTION_KEY must be set");
    assert!(
        operator_mfa_secret.len() >= 32,
        "OPERATOR_MFA_ENCRYPTION_KEY must contain at least 32 bytes"
    );
    let state = SecurityState {
        pool,
        origin: std::env::var("APP_ORIGIN").expect("APP_ORIGIN must be set"),
        cursor_signing_key,
        operator_mfa_key: security::hash(operator_mfa_secret.as_bytes()),
    };
    let app = app(state);
    let listener = tokio::net::TcpListener::bind("0.0.0.0:8080")
        .await
        .expect("API port must be available");
    axum::serve(listener, app).await.expect("API server failed");
}

fn app(state: SecurityState) -> Router {
    let admin_api = Router::new()
        .route("/api/v1/admin/me", get(operators::current_operator))
        .route("/api/v1/admin/reports", get(moderation::list_reports))
        .route(
            "/api/v1/admin/operators",
            axum::routing::post(administration::create_operator),
        )
        .route(
            "/api/v1/admin/operators/{operatorId}/roles",
            axum::routing::put(administration::replace_operator_roles),
        )
        .route(
            "/api/v1/admin/users/{userId}",
            get(administration::get_basic_account),
        )
        .route(
            "/api/v1/admin/users/{userId}/restrictions",
            axum::routing::post(administration::apply_restriction),
        )
        .route("/api/v1/admin/appeals", get(administration::list_appeals))
        .route(
            "/api/v1/admin/appeals/{appealId}",
            axum::routing::patch(administration::review_appeal),
        )
        .route(
            "/api/v1/admin/audit-logs",
            get(administration::read_audit_log),
        )
        .route(
            "/api/v1/admin/auth/mfa/challenge",
            axum::routing::post(operators::confirm_mfa),
        )
        .route(
            "/api/v1/admin/auth/mfa/recovery-codes",
            axum::routing::post(operators::rotate_recovery_codes),
        )
        .route("/api/v1/admin/sessions", get(operators::list_sessions))
        .route(
            "/api/v1/admin/sessions/{sessionId}",
            axum::routing::delete(operators::revoke_session),
        )
        .route(
            "/api/v1/admin/reports/{reportId}",
            get(moderation::get_report).patch(moderation::review_report),
        )
        .route(
            "/api/v1/admin/auth/logout",
            axum::routing::delete(operators::logout),
        )
        .layer(middleware::from_fn_with_state(
            state.clone(),
            operators::require_operator_session,
        ));
    let protected_api = Router::new()
        .route(
            "/users/me",
            get(profile::get_current_user).patch(profile::update_current_user),
        )
        .route("/posts", axum::routing::post(posts::create_post))
        .route(
            "/posts/{postId}",
            get(posts::get_post)
                .patch(posts::update_post)
                .delete(posts::delete_post),
        )
        .route(
            "/posts/{postId}/replies",
            get(posts::list_replies).post(posts::create_reply),
        )
        .route(
            "/posts/{postId}/reports",
            axum::routing::post(reports::create_report),
        )
        .route(
            "/reports/{reportId}",
            get(reports::get_report).patch(reports::update_report),
        )
        .route(
            "/users/{targetUserId}/follow",
            axum::routing::put(social::follow_user).delete(social::unfollow_user),
        )
        .route(
            "/users/{targetUserId}/block",
            axum::routing::post(relationships::block_user).delete(relationships::unblock_user),
        )
        .route(
            "/users/{targetUserId}/mute",
            axum::routing::post(relationships::mute_user).delete(relationships::unmute_user),
        )
        .route("/follow-requests", get(social::list_follow_requests))
        .route(
            "/follow-requests/{relationshipId}/accept",
            axum::routing::post(social::accept_follow_request),
        )
        .route(
            "/follow-requests/{relationshipId}/reject",
            axum::routing::post(social::reject_follow_request),
        )
        .route("/users/{userId}/followers", get(social::list_followers))
        .route("/users/{userId}/following", get(social::list_following))
        .route("/timelines/home", get(social::home_timeline))
        .route(
            "/sessions",
            get(sessions::list_sessions).delete(sessions::revoke_all_sessions),
        )
        .route(
            "/sessions/current",
            axum::routing::delete(sessions::revoke_current_session),
        )
        .route(
            "/sessions/{sessionId}",
            axum::routing::delete(sessions::revoke_session),
        )
        .route(
            "/users/me/deletion-requests",
            axum::routing::post(account_deletion::request_deletion),
        )
        .fallback(api::not_found)
        .layer(middleware::from_fn_with_state(
            state.clone(),
            security::require_session,
        ));
    Router::new()
        .route("/healthz", get(|| async { "ok" }))
        .route(
            "/api/v1/registrations",
            axum::routing::post(registration::register),
        )
        .route(
            "/api/v1/auth/login",
            axum::routing::post(registration::login),
        )
        .route(
            "/api/v1/admin/auth/login",
            axum::routing::post(operators::login),
        )
        .route(
            "/api/v1/admin/auth/mfa/enroll",
            axum::routing::post(operators::enroll_mfa),
        )
        .route(
            "/api/v1/appeals",
            axum::routing::post(administration::create_appeal),
        )
        .route(
            "/api/v1/users/me/deletion-requests/current/cancel",
            axum::routing::post(account_deletion::cancel_deletion),
        )
        .route(
            "/api/v1/users/me/deletion-requests/current/restore",
            axum::routing::post(account_deletion::restore_account),
        )
        .merge(admin_api)
        .route("/readyz", get(ready))
        .route("/openapi.json", get(api::openapi))
        .route("/metrics", get(observability::metrics))
        .nest("/api/v1", protected_api)
        .fallback(api::not_found)
        .layer(
            ServiceBuilder::new()
                .layer(SetRequestIdLayer::new(
                    observability::REQUEST_ID_HEADER,
                    observability::MakeRequestIdFromCSPRNG,
                ))
                .layer(PropagateRequestIdLayer::new(
                    observability::REQUEST_ID_HEADER,
                ))
                .layer(TraceLayer::new_for_http())
                .layer(middleware::from_fn(observability::count_requests)),
        )
        .layer(DefaultBodyLimit::max(16 * 1024))
        .with_state(state)
}

async fn ready(State(state): State<SecurityState>) -> StatusCode {
    match state.pool.acquire().await {
        Ok(_) => StatusCode::OK,
        Err(_) => StatusCode::SERVICE_UNAVAILABLE,
    }
}

#[cfg(test)]
mod tests {
    use super::*;
    use axum::{body::Body, http::Request};
    use sqlx::PgPool;
    use tower::ServiceExt;

    #[tokio::test]
    async fn api_is_deny_by_default() {
        let pool = PgPool::connect_lazy("postgres://miz:miz@localhost/miz").unwrap();
        let response = app(SecurityState {
            pool,
            origin: "https://m1z.jp".to_owned(),
            cursor_signing_key: vec![7; 32],
            operator_mfa_key: [7; 32],
        })
        .oneshot(
            Request::get("/api/v1/unlisted")
                .body(Body::empty())
                .unwrap(),
        )
        .await
        .unwrap();
        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
        let request_id = response.headers()[observability::REQUEST_ID_HEADER]
            .to_str()
            .unwrap();
        assert_eq!(request_id.len(), 22);
    }

    #[tokio::test]
    async fn authenticated_post_creation_uses_the_http_contract() {
        let Ok(database_url) = std::env::var("DATABASE_URL") else {
            return;
        };
        let pool = PgPool::connect(&database_url).await.unwrap();
        infrastructure::migrate(&pool).await.unwrap();
        let user_id = miz_api::domain::UserId::new().unwrap();
        let target_id = miz_api::domain::UserId::new().unwrap();
        let session_id = miz_api::domain::SessionId::new().unwrap();
        let tokens = security::SessionTokens::generate().unwrap();
        sqlx::query(
            "INSERT INTO users (id, display_name) VALUES ($1, 'HTTP Author'), ($2, 'HTTP Target')",
        )
        .bind(user_id.to_bytes().to_vec())
        .bind(target_id.to_bytes().to_vec())
        .execute(&pool)
        .await
        .unwrap();
        sqlx::query(
            "INSERT INTO sessions (id, user_id, token_hash, csrf_token_hash, idle_expires_at, absolute_expires_at) \
             VALUES ($1, $2, $3, $4, now() + INTERVAL '7 days', now() + INTERVAL '30 days')",
        )
        .bind(session_id.to_bytes().to_vec())
        .bind(user_id.to_bytes().to_vec())
        .bind(tokens.session_hash().to_vec())
        .bind(tokens.csrf_hash().to_vec())
        .execute(&pool)
        .await
        .unwrap();

        let app = app(SecurityState {
            pool,
            origin: "https://m1z.jp".to_owned(),
            cursor_signing_key: vec![7; 32],
            operator_mfa_key: [7; 32],
        });
        let response = app
            .clone()
            .oneshot(
                Request::post("/api/v1/posts")
                    .header("content-type", "application/json")
                    .header("origin", "https://m1z.jp")
                    .header("x-csrf-token", &tokens.csrf)
                    .header("idempotency-key", "http-post-test")
                    .header(
                        "cookie",
                        format!(
                            "__Host-miz_session={}; __Host-miz_csrf={}",
                            tokens.session, tokens.csrf
                        ),
                    )
                    .body(Body::from(r#"{"content":"hello\nworld"}"#))
                    .unwrap(),
            )
            .await
            .unwrap();
        assert_eq!(response.status(), StatusCode::CREATED);
        assert_eq!(
            response.headers()[axum::http::header::CONTENT_TYPE],
            "application/json"
        );

        let response = app
            .oneshot(
                Request::put(format!("/api/v1/users/{target_id}/follow"))
                    .header("origin", "https://m1z.jp")
                    .header("x-csrf-token", &tokens.csrf)
                    .header(
                        "cookie",
                        format!(
                            "__Host-miz_session={}; __Host-miz_csrf={}",
                            tokens.session, tokens.csrf
                        ),
                    )
                    .body(Body::empty())
                    .unwrap(),
            )
            .await
            .unwrap();
        assert_eq!(response.status(), StatusCode::OK);
    }
}