# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MIZ API
  version: 0.1.0
  description: The authoritative HTTP contract between the Rust API and MIZ clients.
servers:
  - url: /
security:
  - sessionCookie: []
paths:
...
  /api/v1/registrations:
    post:
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Account created and signed in
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Signed in
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/auth/login:
    post:
      operationId: operatorLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorLoginRequest' }
      responses:
        '200':
          description: Operator signed in with MFA
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/me:
    get:
      operationId: getCurrentOperator
      security:
        - operatorSessionCookie: []
      responses:
        '200':
          description: Current operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/reports:
    get:
      operationId: listAdminReports
      security:
        - operatorSessionCookie: []
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [received, inReview, actioned, noAction], default: received }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Moderation report queue
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminReportList' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/reports/{reportId}:
    get:
      operationId: getAdminReport
      security:
        - operatorSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report with immutable evidence
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminReport' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: reviewAdminReport
      security:
        - operatorSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewReportRequest' }
      responses:
        '200':
          description: Updated report review state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminReport' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/auth/logout:
    delete:
      operationId: operatorLogout
      security:
        - operatorSessionCookie: []
      responses:
        '204': { description: Operator session revoked }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/auth/mfa/enroll:
    post:
      operationId: enrollOperatorMfa
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorEnrollmentRequest' }
      responses:
        '204': { description: MFA enrollment completed }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/auth/mfa/challenge:
    post:
      operationId: confirmOperatorMfa
      security:
        - operatorSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorMfaRequest' }
      responses:
        '204': { description: MFA freshness renewed }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/auth/mfa/recovery-codes:
    post:
      operationId: rotateOperatorRecoveryCodes
      security:
        - operatorSessionCookie: []
      responses:
        '200':
          description: New one-time recovery codes; returned once
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecoveryCodes' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/sessions:
    get:
      operationId: listOperatorSessions
      security:
        - operatorSessionCookie: []
      responses:
        '200':
          description: Active operator sessions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OperatorSession' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/sessions/{sessionId}:
    delete:
      operationId: revokeOperatorSession
      security:
        - operatorSessionCookie: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/ObjectId' }
      responses:
        '204': { description: Operator session revoked }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/operators:
    post:
      operationId: createOperator
      security:
        - operatorSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOperatorRequest' }
      responses:
        '201':
          description: Pending operator and one-time MFA enrollment material
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorEnrollment' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/operators/{operatorId}/roles:
    put:
      operationId: replaceOperatorRoles
      security:
        - operatorSessionCookie: []
      parameters:
        - name: operatorId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/ObjectId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceOperatorRolesRequest' }
      responses:
        '204': { description: Operator roles replaced }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/users/{userId}:
    get:
      operationId: getAdminBasicAccount
      security:
        - operatorSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Basic account information without private content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BasicAccount' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/users/{userId}/restrictions:
    post:
      operationId: applyUserRestriction
      security:
        - operatorSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestrictionRequest' }
      responses:
        '201':
          description: Applied restriction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Restriction' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /api/v1/appeals:
    post:
      operationId: createModerationAppeal
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAppealRequest' }
      responses:
        '201':
          description: Appeal created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Appeal' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/appeals:
    get:
      operationId: listModerationAppeals
      security:
        - operatorSessionCookie: []
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, upheld, overturned], default: pending }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Appeal review queue
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminAppeal' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/appeals/{appealId}:
    patch:
      operationId: reviewModerationAppeal
      security:
        - operatorSessionCookie: []
      parameters:
        - name: appealId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/ObjectId' }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewAppealRequest' }
      responses:
        '200':
          description: Reviewed appeal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Appeal' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/audit-logs:
    get:
      operationId: readAuditLog
      security:
        - operatorSessionCookie: []
      parameters:
        - { name: targetType, in: query, required: true, schema: { type: string, enum: [user, post, report, operator, appeal] } }
        - { name: targetId, in: query, required: true, schema: { $ref: '#/components/schemas/ObjectId' } }
        - { name: from, in: query, required: true, schema: { $ref: '#/components/schemas/Timestamp' } }
        - { name: to, in: query, required: true, schema: { $ref: '#/components/schemas/Timestamp' } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 100 } }
      responses:
        '200':
          description: Time- and subject-bounded audit records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEntry' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/v1/sessions:
    get:
      operationId: listSessions
      responses:
        '200':
          description: Active sessions for the current user
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: revokeAllSessions
      responses:
        '204':
          description: All sessions revoked
          headers:
            Set-Cookie:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/sessions/current:
    delete:
      operationId: revokeCurrentSession
      responses:
        '204':
          description: Current session revoked
          headers:
            Set-Cookie:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/sessions/{sessionId}:
    delete:
      operationId: revokeSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/ObjectId' }
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/users/me:
    get:
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateCurrentUser
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/users/me/deletion-requests:
    post:
      operationId: requestAccountDeletion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestAccountDeletion' }
      responses:
        '201':
          description: Account deletion requested and all sessions revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountDeletionRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/users/me/deletion-requests/current/cancel:
    post:
      operationId: cancelAccountDeletion
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestoreAccountRequest' }
      responses:
        '200':
          description: Pending account deletion cancelled after identity verification
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountDeletionRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/users/me/deletion-requests/current/restore:
    post:
      operationId: restoreAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestoreAccountRequest' }
      responses:
        '200':
          description: Account restored after identity verification
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountDeletionRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/posts:
    post:
      operationId: createPost
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostRequest' }
      responses:
        '201':
          description: Created post
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/follow-requests:
    get:
      operationId: listFollowRequests
      responses:
        '200':
          description: Pending follow requests for the current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FollowRelationshipList' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-miz_session
    operatorSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-miz_operator_session
  parameters:
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [username, password, displayName, birthDate]
      properties:
        username: { $ref: '#/components/schemas/Handle' }
        password: { type: string, minLength: 12, maxLength: 128, format: password }
        displayName: { type: string, minLength: 1, maxLength: 50 }
        birthDate: { type: string, format: date }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { $ref: '#/components/schemas/Handle' }
        password: { type: string, minLength: 12, maxLength: 128, format: password }
...
    User:
      type: object
      required: [id, handle, displayName, bio, privacy, version, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        handle: { $ref: '#/components/schemas/Handle' }
        displayName: { type: string }
        bio: { type: string, maxLength: 160 }
        privacy: { type: string, enum: [public, private] }
        version: { type: integer, format: int64, minimum: 1 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
...
    AdminReportList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminReport' }
...